requirements.txt
emoji==0.4.5
Jinja2==2.9.6
Markdown==2.6.8
MarkupSafe==1.0
olefile==0.44
Pillow==4.2.1
pymongo==3.4.0
web.py==0.38
